+++ site/index.tsx
@@ -1,5 +1,5 @@
-import { Dashboard } from "Pages/Dashboard/Dashboard";
-import { Layout } from "Pages/Layout";
+import { Dashboard } from "pages/Dashboard/Dashboard";
+import { Layout } from "pages/Layout";
 import { injectAnalytics } from "lib/injectAnalytics";
 import m from "mithril";
